Cleanup as per code review
1. Move keyboard to handwired
2. Consolidate led config away from keymaps
3. Add copyright headers
