+++ keyboards/handwired/dygma/raise/keymaps/ansi/readme.md
@@ -1,5 +1,5 @@
 # The ansi keymap for Dygma's Raise
 
 ```sh
-make dygma/raise:ansi
+make handwired/dygma/raise:ansi
 ```
