+++ keyboards/handwired/dygma/raise/keymaps/iso/readme.md
@@ -1,5 +1,5 @@
 # The iso keymap for Dygma's Raise
 
 ```sh
-make dygma/raise:iso
+make handwired/dygma/raise:iso
 ```
