+++ keyboards/handwired/dygma/raise/mcuconf.h
@@ -1,8 +1,8 @@
-/* Copyright 2018-2021 James Laird-Wah, Islam Sharabash
+/* Copyright 2021 Islam Sharabash
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
+ * the Free Software Foundation, either version 2 of the License, or
  * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -11,12 +11,15 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "raise.h"
+#pragma once
 
-void keyboard_post_init_kb(void) {
-    set_all_leds_to(0, 0, 0);
-    keyboard_post_init_user();
-}
+#include_next <mcuconf.h>
+
+#undef STM32_I2C_USE_I2C1
+#define STM32_I2C_USE_I2C1 TRUE
+
+#undef STM32_I2C_BUSY_TIMEOUT
+#define STM32_I2C_BUSY_TIMEOUT 5
