+++ keyboards/handwired/dygma/raise/readme.md
@@ -11,11 +11,11 @@
 
 Make example for this keyboard (after setting up your build environment):
 
-    make dygma/raise:ansi
+    make handwired/dygma/raise:ansi
 
 Flashing example for this keyboard:
 
-    make dygma/raise:ansi:flash
+    make handwired/dygma/raise:ansi:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
